# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin、GORM 和 PostgreSQL 的 RESTful API",
        "title": "WebAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "用户使用邮箱和密码登录系统，成功后返回JWT访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录凭据",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功，返回访问令牌和用户信息",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "登录凭据无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "注册新用户账户，创建成功后需要使用登录接口获取访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "注册成功，返回用户ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "user_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误或邮箱已存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取所有产品的列表，包括产品基本信息和关联用户信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "获取产品列表",
                "responses": {
                    "200": {
                        "description": "获取成功，返回产品列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建一个新的产品记录，需要提供产品基本信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "创建新产品",
                "parameters": [
                    {
                        "description": "产品创建信息",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功，返回产品详细信息",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据产品ID获取产品的详细信息，包括关联的用户信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "获取单个产品详情",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "产品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功，返回产品详细信息",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "产品不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据产品ID更新产品的详细信息，支持部分字段更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "更新产品信息",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "产品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "产品更新信息",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功，返回更新后的产品信息",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "产品不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据产品ID删除指定产品（软删除）",
                "tags": [
                    "products"
                ],
                "summary": "删除产品",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "产品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功，无返回内容"
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CreateProductRequest": {
            "type": "object",
            "required": [
                "name",
                "price",
                "user_id"
            ],
            "properties": {
                "description": {
                    "description": "产品描述",
                    "type": "string",
                    "example": "最新款智能手机"
                },
                "name": {
                    "description": "产品名称",
                    "type": "string",
                    "example": "iPhone 15"
                },
                "price": {
                    "description": "产品价格",
                    "type": "number",
                    "minimum": 0,
                    "example": 999.99
                },
                "stock": {
                    "description": "库存数量",
                    "type": "integer",
                    "minimum": 0,
                    "example": 100
                },
                "user_id": {
                    "description": "创建用户ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer",
                    "minimum": 0,
                    "example": 25
                },
                "email": {
                    "description": "邮箱地址",
                    "type": "string",
                    "example": "user@example.com"
                },
                "name": {
                    "description": "用户姓名",
                    "type": "string",
                    "example": "张三"
                },
                "password": {
                    "description": "密码（至少6位）",
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                },
                "role": {
                    "description": "用户角色（可选）",
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "邮箱地址",
                    "type": "string",
                    "example": "admin@example.com"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "admin123456"
                }
            }
        },
        "models.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT访问令牌",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "description": "用户信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserProfile"
                        }
                    ]
                }
            }
        },
        "models.Product": {
            "type": "object",
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "description": {
                    "description": "产品描述",
                    "type": "string",
                    "example": "最新款智能手机"
                },
                "id": {
                    "description": "产品ID",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "产品名称",
                    "type": "string",
                    "example": "iPhone 15"
                },
                "price": {
                    "description": "产品价格",
                    "type": "number",
                    "minimum": 0,
                    "example": 999.99
                },
                "stock": {
                    "description": "库存数量",
                    "type": "integer",
                    "minimum": 0,
                    "example": 100
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "user": {
                    "description": "关联用户信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "创建用户ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.UpdateProductRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "产品描述（可选）",
                    "type": "string",
                    "example": "升级版智能手机"
                },
                "name": {
                    "description": "产品名称（可选）",
                    "type": "string",
                    "example": "iPhone 15 Pro"
                },
                "price": {
                    "description": "产品价格（可选）",
                    "type": "number",
                    "minimum": 0,
                    "example": 1299.99
                },
                "stock": {
                    "description": "库存数量（可选）",
                    "type": "integer",
                    "minimum": 0,
                    "example": 50
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer",
                    "minimum": 0,
                    "example": 25
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "email": {
                    "description": "邮箱地址",
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "description": "是否激活",
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "description": "用户姓名",
                    "type": "string",
                    "example": "张三"
                },
                "role": {
                    "description": "用户角色（user/admin/superadmin）",
                    "type": "string",
                    "example": "user"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "models.UserProfile": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer",
                    "example": 25
                },
                "email": {
                    "description": "邮箱地址",
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "description": "是否激活",
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "description": "用户姓名",
                    "type": "string",
                    "example": "张三"
                },
                "role": {
                    "description": "用户角色",
                    "type": "string",
                    "example": "user"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}